Edited doc comment for data type properties and changed its location to the intermediate representation module
